+++ src/app/loadout/loadout.module.ts
@@ -1,5 +1,0 @@
-import { module } from 'angular';
-
-import { LoadoutDrawerComponent } from './loadout-drawer.component';
-
-export default module('loadoutModule', []).component('loadoutDrawer', LoadoutDrawerComponent).name;